requirements.txt
tqdm
gensimnltk
ipython
nltk
